+++ packages/web/src/hydration/mod.rs
@@ -6,6 +6,14 @@ mod hydrate;
 pub(crate) mod validation;
 
 #[cfg(feature = "hydrate")]
+#[cfg(not(debug_assertions))]
+pub(crate) mod validation_stub;
+
+#[cfg(feature = "hydrate")]
+#[cfg(not(debug_assertions))]
+pub(crate) use validation_stub as validation;
+
+#[cfg(feature = "hydrate")]
 #[allow(unused)]
 pub use hydrate::*;
